backend/requirements.txt
torch
torchvision
fastapi
uvicorn[standard]
numpy
opencv-python
pillow
matplotlib
tqdm
python-multipart
